requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
pydantic==2.12.0
dotenv==1.0.0
supabase==2.15.1
openai==1.12
